[RayJob][Status][13/n] Make suspend operation atomic by introducing the new status `Suspending` (#1798)

+++ ray-operator/controllers/ray/rayjob_controller.go
@@ -143,6 +143,36 @@ func (r *RayJobReconciler) Reconcile(ctx context.Context, request ctrl.Request) 
 			return ctrl.Result{RequeueAfter: RayJobDefaultRequeueDuration}, nil
 		}
 		// TODO (kevin85421): Don't return here; we still need to move some of the code below into this block.
+	case rayv1.JobDeploymentStatusSuspending:
+		// The `suspend` operation should be atomic. In other words, if users set the `suspend` flag to true and then immediately
+		// set it back to false, either all of the RayJob's associated resources should be cleaned up, or no resources should be
+		// cleaned up at all. To keep the atomicity, if a RayJob is in the `Suspending` status, we should delete all of its
+		// associated resources and then transition the status to `Suspended` no matter the value of the `suspend` flag.
+
+		// TODO (kevin85421): Currently, Ray doesn't have a best practice to stop a Ray job gracefully. At this moment,
+		// KubeRay doesn't stop the Ray job before suspending the RayJob. If users want to stop the Ray job by SIGTERM,
+		// users need to set the Pod's preStop hook by themselves.
+		isReleaseComplete, err := r.releaseComputeResources(ctx, rayJobInstance, true)
+		if err != nil {
+			return ctrl.Result{RequeueAfter: RayJobDefaultRequeueDuration}, err
+		}
+		if !isReleaseComplete {
+			r.Log.Info("The release of the compute resources has not been completed yet. " +
+				"Wait for the resources to be deleted before the status transitions to avoid a resource leak.")
+			return ctrl.Result{RequeueAfter: RayJobDefaultRequeueDuration}, nil
+		}
+
+		// Reset the RayCluster and Ray job related status.
+		rayJobInstance.Status.RayClusterStatus = rayv1.RayClusterStatus{}
+		rayJobInstance.Status.RayClusterName = ""
+		rayJobInstance.Status.DashboardURL = ""
+		rayJobInstance.Status.JobId = ""
+		rayJobInstance.Status.Message = ""
+
+		if err = r.updateState(ctx, rayJobInstance, nil, rayv1.JobStatusNew, rayv1.JobDeploymentStatusSuspended); err != nil {
+			return ctrl.Result{RequeueAfter: RayJobDefaultRequeueDuration}, err
+		}
+		return ctrl.Result{RequeueAfter: RayJobDefaultRequeueDuration}, nil
 	case rayv1.JobDeploymentStatusSuspended:
 		if !rayJobInstance.Spec.Suspend {
 			r.Log.Info("The status is 'Suspended', but the suspend flag is false. Transition the status to 'New'.")
@@ -173,7 +203,7 @@ func (r *RayJobReconciler) Reconcile(ctx context.Context, request ctrl.Request) 
 				r.Log.Info(fmt.Sprintf("shutdownTime not reached, requeue this RayJob for %d seconds", delta))
 				return ctrl.Result{RequeueAfter: time.Duration(delta) * time.Second}, nil
 			} else {
-				if err = r.releaseComputeResources(ctx, rayJobInstance, false); err != nil {
+				if _, err = r.releaseComputeResources(ctx, rayJobInstance, false); err != nil {
 					return ctrl.Result{RequeueAfter: RayJobDefaultRequeueDuration}, err
 				}
 			}
@@ -188,6 +218,9 @@ func (r *RayJobReconciler) Reconcile(ctx context.Context, request ctrl.Request) 
 	}
 	// If there is no cluster instance and no error suspend the job deployment
 	if rayClusterInstance == nil {
+		// TODO (kevin85421): Note that if we suspend the RayJob in the `Initializing` status, the RayJob will not
+		// transition to the `Suspending` status. It will transition to the `Suspended` status directly. We should
+		// unify the code path.
 		err = r.updateState(ctx, rayJobInstance, nil, rayJobInstance.Status.JobStatus, rayv1.JobDeploymentStatusSuspended)
 		if err != nil {
 			return ctrl.Result{RequeueAfter: RayJobDefaultRequeueDuration}, err
@@ -244,39 +277,15 @@ func (r *RayJobReconciler) Reconcile(ctx context.Context, request ctrl.Request) 
 		return ctrl.Result{RequeueAfter: RayJobDefaultRequeueDuration}, err
 	}
 
+	// TODO (kevin85421): This is unnecessary because the status has to be Running if we have reached here.
 	if rayJobInstance.Status.JobDeploymentStatus == rayv1.JobDeploymentStatusRunning {
 		// If suspend flag is set AND
 		// the RayJob is submitted against the RayCluster created by THIS job, then
 		// try to gracefully stop the Ray job and delete (suspend) the cluster
 		if rayJobInstance.Spec.Suspend && len(rayJobInstance.Spec.ClusterSelector) == 0 {
-			if !rayv1.IsJobTerminal(jobInfo.JobStatus) {
-				err := rayDashboardClient.StopJob(ctx, rayJobInstance.Status.JobId, &r.Log)
-				if err != nil {
-					return ctrl.Result{RequeueAfter: RayJobDefaultRequeueDuration}, err
-				}
-			}
-			if jobInfo.JobStatus != rayv1.JobStatusStopped {
-				return ctrl.Result{RequeueAfter: RayJobDefaultRequeueDuration}, nil
-			}
-
-			if err = r.releaseComputeResources(ctx, rayJobInstance, true); err != nil {
-				return ctrl.Result{RequeueAfter: RayJobDefaultRequeueDuration}, err
-			}
-			// Since RayCluster instance is gone, remove it status also
-			// on RayJob resource
-			rayJobInstance.Status.RayClusterStatus = rayv1.RayClusterStatus{}
-			rayJobInstance.Status.RayClusterName = ""
-			rayJobInstance.Status.DashboardURL = ""
-			rayJobInstance.Status.JobId = ""
-			rayJobInstance.Status.Message = ""
-			err = r.updateState(ctx, rayJobInstance, jobInfo, rayv1.JobStatusStopped, rayv1.JobDeploymentStatusSuspended)
-			if err != nil {
-				return ctrl.Result{RequeueAfter: RayJobDefaultRequeueDuration}, err
-			}
-			r.Log.Info("rayJob suspended", "RayJob", rayJobInstance.Name)
-			r.Recorder.Eventf(rayJobInstance, corev1.EventTypeNormal, "Suspended", "Suspended RayJob %s", rayJobInstance.Name)
-			return ctrl.Result{RequeueAfter: RayJobDefaultRequeueDuration}, nil
-			// Job may takes long time to start and finish, let's just periodically requeue the job and check status.
+			r.Log.Info("Try to transition the status to `Suspending`", "RayJob", rayJobInstance.Name)
+			err = r.updateState(ctx, rayJobInstance, nil, rayJobInstance.Status.JobStatus, rayv1.JobDeploymentStatusSuspending)
+			return ctrl.Result{RequeueAfter: RayJobDefaultRequeueDuration}, err
 		}
 	}
 
@@ -389,26 +398,30 @@ func (r *RayJobReconciler) createNewK8sJob(ctx context.Context, rayJobInstance *
 // In the future, we may also need to delete other Kubernetes resources. Note that
 // this function doesn't delete the Kubernetes Job. Instead, we use the built-in
 // TTL mechanism of the Kubernetes Job for deletion.
-func (r *RayJobReconciler) releaseComputeResources(ctx context.Context, rayJobInstance *rayv1.RayJob, isSuspend bool) error {
+func (r *RayJobReconciler) releaseComputeResources(ctx context.Context, rayJobInstance *rayv1.RayJob, isSuspend bool) (bool, error) {
 	namespace := rayJobInstance.Namespace
 	clusterIdentifier := types.NamespacedName{
 		Name:      rayJobInstance.Status.RayClusterName,
 		Namespace: namespace,
 	}
+	isClusterNotFound, isJobNotFound := false, !isSuspend
+
 	cluster := rayv1.RayCluster{}
 	if err := r.Get(ctx, clusterIdentifier, &cluster); err != nil {
-		if !errors.IsNotFound(err) {
-			return err
+		if errors.IsNotFound(err) {
+			// If the cluster is not found, it means the cluster has been already deleted.
+			// Don't return error to make this function idempotent.
+			isClusterNotFound = true
+			r.Log.Info("The associated cluster has been already deleted and it can not be found", "RayCluster", clusterIdentifier)
+		} else {
+			return false, err
 		}
-		// If the cluster is not found, it means the cluster has been already deleted.
-		// Don't return error to make this function idempotent.
-		r.Log.Info("The associated cluster has been already deleted and it can not be found", "RayCluster", clusterIdentifier)
 	} else {
-		if cluster.DeletionTimestamp != nil {
+		if !cluster.DeletionTimestamp.IsZero() {
 			r.Log.Info("The cluster deletion is ongoing.", "rayjob", rayJobInstance.Name, "raycluster", cluster.Name)
 		} else {
 			if err := r.Delete(ctx, &cluster); err != nil {
-				return err
+				return false, err
 			}
 			r.Log.Info("The associated cluster is deleted", "RayCluster", clusterIdentifier)
 			r.Recorder.Eventf(rayJobInstance, corev1.EventTypeNormal, "Deleted", "Deleted cluster %s", rayJobInstance.Status.RayClusterName)
@@ -423,19 +436,29 @@ func (r *RayJobReconciler) releaseComputeResources(ctx context.Context, rayJobIn
 		job := &batchv1.Job{}
 		if err := r.Client.Get(ctx, client.ObjectKey{Namespace: namespace, Name: jobName}, job); err != nil {
 			if errors.IsNotFound(err) {
+				isJobNotFound = true
 				r.Log.Info("The submitter Kubernetes Job has been already deleted", "RayJob", rayJobInstance.Name, "Kubernetes Job", job.Name)
-				return nil
 			} else {
 				r.Log.Error(err, "Failed to get Kubernetes Job")
-				return err
+				return false, err
+			}
+		} else {
+			if !job.DeletionTimestamp.IsZero() {
+				r.Log.Info("The Job deletion is ongoing.", "RayJob", rayJobInstance.Name, "Submitter K8s Job", job.Name)
+			} else {
+				if err := r.Client.Delete(ctx, job, client.PropagationPolicy(metav1.DeletePropagationBackground)); err != nil {
+					return false, err
+				}
+				r.Log.Info("The associated submitter Job is deleted", "RayJob", rayJobInstance.Name, "Submitter K8s Job", job.Name)
+				r.Recorder.Eventf(rayJobInstance, corev1.EventTypeNormal, "Deleted", "Deleted submitter K8s Job %s", job.Name)
 			}
 		}
-		if err := r.Client.Delete(ctx, job, client.PropagationPolicy(metav1.DeletePropagationBackground)); err != nil {
-			r.Log.Error(err, "Failed to delete the Kubernetes Job")
-			return err
-		}
 	}
-	return nil
+
+	isReleaseComplete := isClusterNotFound && isJobNotFound
+	r.Log.Info("releaseComputeResources", "isClusterNotFound", isClusterNotFound, "isJobNotFound", isJobNotFound, "isReleaseComplete", isReleaseComplete)
+
+	return isReleaseComplete, nil
 }
 
 // isJobSucceedOrFail indicates whether the job comes into end status.
